only works for first column, have temporary matrix, gotta fix this in the 2nd iter

+++ matrix_functions.c
@@ -2,12 +2,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int* create_matrix(int numRows, int numCols){
+float* create_matrix(int numRows, int numCols){
     if (numRows <= 0 || numCols <= 0){
         fprintf(stderr, "invalid inputs\n");
         return NULL;
     }
-    int* matrix = (int*)malloc(numRows*numCols*sizeof(int));
+    float* matrix = (float*)malloc(numRows*numCols*sizeof(float));
     if (matrix == NULL){
         fprintf(stderr, "memory allocation failed");
         exit(1);
@@ -15,11 +15,11 @@ int* create_matrix(int numRows, int numCols){
     return matrix;
 }
 
-void initiate_matrix(int numRows, int numCols, int* matrix){
+void initiate_matrix(int numRows, int numCols, float* matrix){
     printf("insert the matrix: ");
     for(int i = 0; i < numRows; i++){
         for(int j = 0; j < numCols; j++){
-            if(scanf("%d", &matrix[i*numCols + j]) != 1){
+            if(scanf("%f", &matrix[i*numCols + j]) != 1){
                 fprintf(stderr, "invalid inputs\n");
                 return;
             }
@@ -27,15 +27,15 @@ void initiate_matrix(int numRows, int numCols, int* matrix){
     }
 }
 
-void free_matrix(int* matrix){
+void free_matrix(float* matrix){
     free(matrix);
 }
 
-void print_matrix(int numRows, int numCols, int* matrix){
+void print_matrix(int numRows, int numCols, float* matrix){
     printf("the matrix: \n");
     for(int i = 0; i < numRows; i++){
         for(int j = 0; j < numCols; j++){
-            printf("%d\t", matrix[i*numCols + j]);
+            printf("%f\t", matrix[i*numCols + j]);
         }
         printf("\n");
     }
